Update Linkables.xml
This is a separate change from my xml selector update. This is more of a 'personal preference' patch for the more linkables support, which changes the tier 1 ML toolbox(the blue one) to use the art assets for the vanilla(green) toolbox, since it is no longer being used in favor of the Gloomy one. It also makes the tier 2 toolbox(the red one) appear to be a brighter shade of red(again, using the vanilla textures) to more closely resemble an actual IRL red mechanics toolbox. https://files.catbox.moe/xz385r.PNG